LayoutTests/editing/inserting/6633727-expected.txt

This tests the fix for <rdar://problem/6633727> Hitting return at the end of a line with an anchor jumps me to the bottom of the message. If the test has passed, the numbers should be in order, and only "1" should be a link.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS editable.innerHTML is '<a href="#" id="anchor">1</a><div><a href="#" id="anchor"></a>2<br><div>3</div></div>'
PASS sel.baseNode is editable.childNodes.item(1).childNodes.item(1)
PASS sel.baseNode.nodeType is Node.TEXT_NODE
PASS sel.baseOffset is 1
PASS successfullyParsed is true

TEST COMPLETE
1
2
3
